+++ .babelrc
@@ -1,4 +1,3 @@
 {
-  "presets": ["env"],
-  "plugins": ["babel-plugin-inline-import"]
+  "presets": ["env"]
 }
