--- example/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["env"],
-  "plugins": ["babel-plugin-inline-import"]
-}
