--- example/dev.js
@@ -1,3 +1,0 @@
-// only for development
-require('babel-register'); // eslint-disable-line
-require('./src/index');
