+++ example/fakeData.js
@@ -1,14 +1,6 @@
-const data = [
+module.exports = [
   { personId: 1, name: 'Kevin Odling' },
   { personId: 2, name: 'Farah Bennett' },
   { personId: 3, name: 'Arian Guthrie' },
   { personId: 4, name: 'Serena Monroe' },
 ];
-
-const resolvers = {
-  Query: {
-    persons: () => data,
-  },
-};
-
-export default resolvers;
