--- example/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "uid-example",
-  "version": "0.1.0",
-  "description": "Example usage of uid directive",
-  "main": "index.js",
-  "license": "MIT",
-  "scripts": {
-    "start": "node dev.js",
-    "test": "jest"
-  },
-  "dependencies": {
-    "apollo-server-express": "^1.3.2",
-    "body-parser": "^1.18.2",
-    "express": "^4.16.3",
-    "graphql": "^0.13.2",
-    "graphql-tools": "^2.23.1",
-    "graphql-directive-uid": "file:../"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.2.2",
-    "babel-jest": "^22.4.3",
-    "babel-plugin-inline-import": "^2.0.6",
-    "babel-preset-env": "^1.6.1",
-    "babel-register": "^6.26.0"
-  }
-}
