+++ package.json
@@ -7,32 +7,24 @@
   "author": "Czysty <lukasz.czyszczonik@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "build": "babel index.js -d dist",
-    "test": "jest",
-    "lint": "eslint .",
-    "prepare": "yarn build"
+    "dev": "nodemon example/index.js",
+    "test": "jest"
   },
   "peerDependencies": {
     "graphql": "^0.13.2",
     "graphql-tools": "^2.23.1"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.2.2",
-    "babel-jest": "^22.4.3",
-    "babel-plugin-inline-import": "^2.0.6",
-    "babel-preset-env": "^1.6.1",
-    "babel-register": "^6.26.0",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-jest": "^21.15.0",
-    "eslint-plugin-prettier": "^2.6.0",
+    "apollo-server-express": "^1.3.4",
+    "body-parser": "^1.18.2",
+    "express": "^4.16.3",
     "graphql": "^0.13.2",
     "graphql-tools": "^2.23.1",
-    "jest": "^22.4.3",
-    "prettier": "^1.11.1",
-    "prettier-eslint": "^8.8.1"
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-polyfill": "^6.26.0",
+    "eslint": "^4.19.1",
+    "eslint-config-callstack-io": "^1.1.1",
+    "jest": "^22.4.3"
   }
 }
